// src/graphql/generated/API.ts
export type CreateLibraryInput = {
  id?: string | null,
  name: string,
  username?: string | null,
  createdAt?: string | null,
  updatedAt?: string | null,
};

export type ModelLibraryConditionInput = {
  name?: ModelStringInput | null,
  createdAt?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
  and?: Array< ModelLibraryConditionInput | null > | null,
  or?: Array< ModelLibraryConditionInput | null > | null,
  not?: ModelLibraryConditionInput | null,
};

export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type Library = {
  __typename: "Library",
  id?: string,
  name?: string,
  dictionaries?: ModelDictionaryConnection,
  username?: string | null,
  createdAt?: string | null,
  updatedAt?: string | null,
};

export type ModelDictionaryConnection = {
  __typename: "ModelDictionaryConnection",
  items?:  Array<Dictionary | null > | null,
  nextToken?: string | null,
};

export type Dictionary = {
  __typename: "Dictionary",
  id?: string,
  title?: string,
  libraryId?: string,
  library?: Library,
  username?: string | null,
  createdAt?: string | null,
  updatedAt?: string | null,
};

export type UpdateLibraryInput = {
  id: string,
  name?: string | null,
  username?: string | null,
  createdAt?: string | null,
  updatedAt?: string | null,
};

export type DeleteLibraryInput = {
  id: string,
};
// ...
export type ModelIDInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};

export type UpdateDictionaryInput = {
  id: string,
  title?: string | null,
  libraryId?: string | null,
  username?: string | null,
  createdAt?: string | null,
  updatedAt?: string | null,
};

export type DeleteDictionaryInput = {
  id: string,
};

export type ModelLibraryFilterInput = {
  id?: ModelIDInput | null,
  name?: ModelStringInput | null,
  username?: ModelStringInput | null,
  createdAt?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
  and?: Array< ModelLibraryFilterInput | null > | null,
  or?: Array< ModelLibraryFilterInput | null > | null,
  not?: ModelLibraryFilterInput | null,
};

export type ModelLibraryConnection = {
  __typename: "ModelLibraryConnection",
  items?:  Array<Library | null > | null,
  nextToken?: string | null,
};
// ...
export type CreateLibraryMutationVariables = {
  input?: CreateLibraryInput,
  condition?: ModelLibraryConditionInput | null,
};

export type CreateLibraryMutation = {
  createLibrary?:  {
    __typename: "Library",
    id: string,
    name: string,
    dictionaries?:  {
      __typename: "ModelDictionaryConnection",
      nextToken?: string | null,
    } | null,
    username?: string | null,
    createdAt?: string | null,
    updatedAt?: string | null,
  } | null,
};
// ...
export type DeleteLibraryMutationVariables = {
  input?: DeleteLibraryInput,
  condition?: ModelLibraryConditionInput | null,
};

export type DeleteLibraryMutation = {
  deleteLibrary?:  {
    __typename: "Library",
    id: string,
    name: string,
    dictionaries?:  {
      __typename: "ModelDictionaryConnection",
      nextToken?: string | null,
    } | null,
    username?: string | null,
    createdAt?: string | null,
    updatedAt?: string | null,
  } | null,
};
// ...
export type ListLibrarysQueryVariables = {
  filter?: ModelLibraryFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListLibrarysQuery = {
  listLibrarys?:  {
    __typename: "ModelLibraryConnection",
    items?:  Array< {
      __typename: "Library",
      id: string,
      name: string,
      username?: string | null,
      createdAt?: string | null,
      updatedAt?: string | null,
    } | null > | null,
    nextToken?: string | null,
  } | null,
};